add spinner when lazy load
